+++ ide-gen-proto.sh
@@ -68,7 +68,7 @@ gen_proto_ts() {
 gen_proto_dotnet accounting
 # gen_proto_java ad
 gen_proto_dotnet cart
-gen_proto_go checkoutservice
+gen_proto_go checkout
 # gen_proto_cpp currency
 # gen_proto_ruby email
 gen_proto_ts frontend